requirements.txt
numpy==1.19.5
opencv-python>=4.5.3
pillow==8.2.0
torch==1.7.1
torchvision==0.8.2